Adding the INCLUDE directory.
It is required for all the reduced Norne cases.

INCLUDE/GRID/ACTNUM_0704.prop was modified to remove one small isolated
region in the mesh. It is not something necessary for the OPM running.
The reason is that it is necessary for MRST running. In default, MRST
can only handle one connected mesh.

And also we keep the original ACTNUM file with name
ACTNUM_0704.prop.original.