Adding the first CASE for the CASE5 series.
The case5 series include the initialization with EQUIL and includes
MULTZ keywords.

This DATA file is a three phase case with live oil and live gas.
Only one property region.

It contains 2 water injections wells and 3 productions wells.
One injection well is with 3 perforations and all the other wells are
single perforation.

One injection well is RATE control and all the other wells are with BHP
control.