update the reference solutions for SPE5CASE1
this changed because of OPM/ewoms#253. strangly enough the jenkins
update of the reference data did not work (#290 was empty) but on my
machine it started to fail locally.

needless to mention that I've looked at the new results using "eyball
norm". I could not see any difference in the curves.