update the reference solution for SPE5CASE1
this is (probably) required because the tolerance to which the
hydrostatic initial condition is solved for in the new "native ebos"
code is four orders of magnitude lower than in the opm-core variant.

As usual, I compared the new summary files with the old ones using
"eyeball norm" and did not see any difference.